Move the abstract componsnt into the parent namespace. Also, give it a new name.
